move tologgerprintstream
